Imported the swipeable view, and fixed some code so it would work within the project.
15 test users and their house images have been added to their respective place in Firebase (Database/Storage) to test the card swiping with real data behind it and also the matching.

I've just started on the matching. I still need to write a function that will go and check if both users have matched with each other. Right now, the only thing that happens is the app is observing their 'matching' endpoint for changes, but it doesn't do anything besides print to the console.

This is a large commit, but I didn't want to commit until I had things working, and I cleaned up some of the code. (It's still pretty messy though)