Renter creation in both CoreData and Firebase now works.
Uploading images to Firebase Storage has also been optimized so that the images get compressed down to the size we specify in code. (This may not be the most efficient way to do it. I think it will work fine, but I'm going to check with my boss at DevMountain to be sure.)